eloapp-13 Wydzielenie RemoveGameSvc
